backend/django-appointment/requirements.txt
asgiref==3.9.1
babel==2.17.0
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.1.7
django-appconf==1.1.0
django-cors-headers==4.9.0
django-guardian==3.2.0
django-phonenumber-field==8.1.0
django-picklefield==3.3
django-q2==1.8.0
django-select2==8.4.1
icalendar==6.3.1
idna==3.10
iniconfig==2.3.0
msgpack==1.1.2
packaging==25.0
phonenumbers==9.0.12
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
